backend/requirements.txt
fastapi
uvicorn
torch
transformers
opencv-python
numpy
pillow
scipy
python-multipart